requirements.txt
argparse
python-dotenv
hvac
requests
pyspark
pytest